long longs 😢

+++ Silver/Contests/CF Round 1076 (Div. 3)/D.cpp
@@ -29,7 +29,7 @@ int main() {
         for (auto& i : b) cin >> i;
 
         sort(a.begin(), a.end());   // l -> g
-        vector<int> pref(N+1, 0);
+        vector<long long> pref(N+1, 0); // this needs to be long longs!!!
         for (int i{0}; i < N; ++i) {
             pref[i+1] = pref[i]+b[i];
         }
@@ -37,7 +37,7 @@ int main() {
         long long ans = 0;
         for (int i{0}; i < N; ++i) {
             long long x = a[i];
-            int swords = N-i;   // # of valid swords
+            long long swords = N-i;   // # of valid swords
             int level = upper_bound(pref.begin(), pref.end(), swords) - pref.begin();
             level--;
             ans = max(ans, x*level);
